+++ packages/ui/src/components/Panel.svelte
@@ -14,7 +14,11 @@
 -->
 <script lang="ts">
   import { createEventDispatcher } from 'svelte'
-  import { Button, IconClose, IconDetails, IconScale, IconScaleFull } from '..'
+  import IconClose from './icons/Close.svelte'
+  import IconDetails from './icons/Details.svelte'
+  import IconScale from './icons/Scale.svelte'
+  import IconScaleFull from './icons/ScaleFull.svelte'
+  import Button from './Button.svelte'
 
   export let innerWidth: number = 0
   export let panelWidth: number = 0
