+++ packages/ui/src/components/calendar/DatePicker.svelte
@@ -15,7 +15,8 @@
 <script lang="ts">
   import { createEventDispatcher } from 'svelte'
   import type { IntlString } from '@anticrm/platform'
-  import ui, { Label } from '../..'
+  import ui from '../../plugin'
+  import Label from '../Label.svelte'
   import DatePresenter from './DatePresenter.svelte'
 
   export let title: IntlString
