+++ packages/ui/src/components/calendar/DateRangePresenter.svelte
@@ -16,8 +16,11 @@
   import type { IntlString } from '@anticrm/platform'
   import { createEventDispatcher, afterUpdate } from 'svelte'
   import { daysInMonth, getMonthName } from './internal/DateUtils'
-  import ui, { Label, Icon, IconClose } from '../..'
+  import ui from '../../plugin'
   import { dpstore, closeDatePopup } from '../../popups'
+  import Label from '../Label.svelte'
+  import Icon from '../Icon.svelte'
+  import IconClose from '../icons/Close.svelte'
   import DPCalendar from './icons/DPCalendar.svelte'
   import DPCalendarOver from './icons/DPCalendarOver.svelte'
   import DateRangePopup from './DateRangePopup.svelte'
