+++ packages/ui/src/components/calendar/Month.svelte
@@ -14,9 +14,19 @@
 -->
 <script lang="ts">
   import { afterUpdate, createEventDispatcher } from 'svelte'
-  import { IconNavPrev, IconNavNext, Icon } from '../..'
-  import { TCellStyle, ICell } from './internal/DateUtils'
-  import { firstDay, day, getWeekDayName, areDatesEqual, getMonthName, daysInMonth } from './internal/DateUtils'
+  import IconNavPrev from '../icons/NavPrev.svelte'
+  import IconNavNext from '../icons/NavNext.svelte'
+  import Icon from '../Icon.svelte'
+  import {
+    firstDay,
+    day,
+    getWeekDayName,
+    areDatesEqual,
+    getMonthName,
+    daysInMonth,
+    TCellStyle,
+    ICell
+  } from './internal/DateUtils'
 
   export let currentDate: Date | null
   export let mondayStart: boolean = true
