+++ packages/ui/src/location.ts
@@ -13,8 +13,8 @@
 // limitations under the License.
 //
 
-import { Location as PlatformLocation } from './types'
 import { writable, derived } from 'svelte/store'
+import { Location as PlatformLocation } from './types'
 
 export function locationToUrl (location: PlatformLocation): string {
   let result = '/'
